requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
